2023 12 22 17:45 하루 일과 종료, 그룹 매칭 기능 구현중 - UI 부분 구현 완료
